Rewrote animation renderer + massive cleanup
* Moved to Qt 5.15.
* Moved to Mingw 8.1 64bits.
* Now using C++17 standard(in theory).
* Rewrote the animation renderer to be less of a buggy mess. The animated sprite item is now simply pre-rendered, and re-rendered when needed. This means, a whole lot less potential bugs, and segfaults..
* Moved the tabs using the animation renderer to their own class. Ideally all tabs will move to their own class, in order to keep the main window code clean.
* Fixed all the deprecation warning, reorder warning, and etc..
* Removed a lot of redundant calls to std::move.. But, I'm getting some doubt that the compiler is actually properly handling copy ellision..
* Renamed a lot of files to have a more coherent naming convention.
* Moved a lot of code to separate files.
* Re-structured directory layout a bit. More to be done.
- Lots left to be done!

+++ src/ppmdu/fmts/packfile.hpp
@@ -113,7 +113,7 @@ namespace fmt
             if( sz > std::numeric_limits<uint32_t>::max() )
                 throw std::overflow_error("PackFileWriter::AppendSubFile(): File is longer than a 32 bit unsigned integer can contain!");
 
-            packfile_hdr::fentry curentry{ m_data.size(), static_cast<uint32_t>(sz) };
+            packfile_hdr::fentry curentry{ static_cast<uint32_t>(m_data.size()), static_cast<uint32_t>(sz) };
 
             itput = std::copy( beg, end, itput );
 
@@ -130,7 +130,7 @@ namespace fmt
             std::lock_guard<std::mutex> lk(m_mtx);
             //Calculate Header + toc + padding len
             size_t hdrlen       = 8 + ((m_hdr.tocentries.size() + 1) * packfile_hdr::SzFEntry);
-            size_t hdrandpadlen = utils::CalculatePaddedLengthTotal(hdrlen, AlignHeaderOn);
+            size_t hdrandpadlen = utils::CalculatePaddedLengthTotal(hdrlen, static_cast<size_t>(AlignHeaderOn));
 
             //Write Leading 0
             where = utils::writeBytesFrom( static_cast<uint32_t>(0), where );
